backend/pyproject.toml
[project]
name = "iam-lifecycle-demo"
version = "0.1.0"
description = "IAM lifecycle automation demo (Google + GitHub)"
authors = [{ name = "Everyday Identity" }]
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.112",
  "uvicorn[standard]>=0.30",
  "pydantic[email]>=2.8",
  "SQLAlchemy>=2.0",
  "psycopg[binary,pool]>=3.2",
  "alembic>=1.13",
  "httpx>=0.27",
  "python-dotenv>=1.0",
  "celery>=5.4",
  "redis>=5.0",
  "python-multipart>=0.0.9",
  "PyYAML>=6.0",
  "PyJWT>=2.8.0",
  "python-jose[cryptography]>=3.3.0",
  "bcrypt>=4.1.0",
  "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "pytest-asyncio>=0.23",
  "ruff>=0.5",
  "mypy>=1.10",
]

[tool.ruff]
line-length = 100
